+++ QuickStackStore/Source/Modules/TrashModule.cs
@@ -87,13 +87,27 @@ namespace QuickStackStore
         [HarmonyPatch(typeof(InventoryGui))]
         internal static class TrashItemsPatches
         {
+            internal static bool hasOpenedInventoryOnce = false;
+
             // slightly lower priority so we get rendered on top of equipment slot mods
             // (lower priority -> later rendering -> you get rendered on top)
             [HarmonyPriority(Priority.LowerThanNormal)]
             [HarmonyPatch(nameof(InventoryGui.Show))]
             [HarmonyPostfix]
-            public static void Show_Postfix(InventoryGui __instance)
+            private static void Show_Postfix(InventoryGui __instance)
             {
+                hasOpenedInventoryOnce = true;
+
+                UpdateTrashCanUI(__instance);
+            }
+
+            internal static void UpdateTrashCanUI(InventoryGui __instance)
+            {
+                if (!hasOpenedInventoryOnce)
+                {
+                    return;
+                }
+
                 if (__instance != InventoryGui.instance)
                 {
                     return;
@@ -259,6 +273,8 @@ namespace QuickStackStore
                 Button button = buttonGo.AddComponent<Button>();
                 button.onClick.AddListener(() => TrashOrTrashFlagItem());
 
+                ControllerButtonHintHelper.AddControllerTooltipToTrashCan(button, tArmor);
+
                 // Add invisible image so we become clickable
                 var image = buttonGo.AddComponent<Image>();
                 image.color = new Color(0, 0, 0, 0);
@@ -271,13 +287,7 @@ namespace QuickStackStore
                 newFrame.GetComponent<RectTransform>().sizeDelta = new Vector2(-8, 22);
                 newFrame.GetComponent<RectTransform>().anchoredPosition = new Vector2(6, 7.5f);
 
-                // Add inventory screen tab
-                UIGroupHandler handler = gameObject.AddComponent<UIGroupHandler>();
-                handler.m_groupPriority = 1;
-                handler.m_enableWhenActiveAndGamepad = newFrame;
-                InventoryGui.instance.m_uiGroups = InventoryGui.instance.m_uiGroups.AddToArray(handler);
-
-                gameObject.AddComponent<TrashHandler>();
+                gameObject.AddComponent<TrashFrameHandler>().frame = newFrame;
             }
 
             protected void Start()
@@ -414,23 +424,13 @@ namespace QuickStackStore
         }
     }
 
-    public class TrashHandler : MonoBehaviour
+    public class TrashFrameHandler : MonoBehaviour
     {
-        private UIGroupHandler handler;
-
-        protected void Awake()
-        {
-            handler = this.GetComponent<UIGroupHandler>();
-        }
+        internal GameObject frame;
 
         protected void Update()
         {
-            if (ZInput.GetButtonDown("JoyButtonA") && handler.IsActive())
-            {
-                TrashModule.TrashOrTrashFlagItem();
-                // Switch back to inventory tab
-                InventoryGui.instance.SetActiveGroup(1);
-            }
+            frame.SetActive(ZInput.IsGamepadActive() && InventoryGui.instance.m_activeGroup == 1);
         }
     }
 }